mkinitramfs: add dropbear configuration.
dropbear needs a password file, passwd-lookup libraries,
etc. It also needs a host key.

With this, it is sufficient to configure the network and
run "dropbear", and it is possible to 'ssh' in to the device
as root using "GnuBee" as the password.

If an "authorized_keys" file is placed in a "dropbear" directory
in the build area, it will be copied to the initramfs and password-less
login will work.

Signed-off-by: NeilBrown <neil@brown.name>