init: allow button to enable ssh-to-initramfs
If the black button is held during boot, the initramfs
will configure the network and run "dropbear" for ssh login.

This can be used for rescue if you mess up your root filesystem,
or to install debian using the "config" script.

The "black" RJ45 (near power plug) is 192.168.10.1
The "blue" RF45 (next one along) gets an address with dhcp.

Signed-off-by: NeilBrown <neil@brown.name>