Update for new 5.1 kernel
Linux 5.1 contains good support for the MT7621 network,
including the integrated switch.
So I've added support for that.

With 5.1 the switch ports have names:

  ethblack ethblue

instead of being vlan ports

  eth0.1 eth0.2

This required some code changes to use the correct interface on the
different kernels, and particularly needed some modification
to the root filesystem to ensure it boots properly.
This changes are made in a tmpfs, they aren't permanent.

Signed-off-by: NeilBrown <neil@brown.name>