mkinitramfs: cope better with swconfig missing.
On mainline kernels. swconfig will not be built
and so will not be installed.
So we should test for its existance before using
it to avoid unsightly error messages.

Separately: don't check for dsa in the network scripts.
I must have been confused when I put that in, but
the current form means that ethblack is never configured.

Signed-off-by: NeilBrown <neil@brown.name>