Remove support for swconfig.
Old 4.4 kernels needed out-of-tree patches to support the ethernet
switch, and that made use of a the 'swconfig' tool.
We have progressed well beyond that point now and there is no longer
a good reason to use such an old kernel, and I'm no longer interested
in maintaining that support.

So remove swconfig and all related support.

Signed-off-by: NeilBrown <neil@brown.name>