Support building smaller firmware for u-boot hacking.
 gbmake firmware gbpc1u-5.6

will create a smaller firmware image - least than 16MB - which is
configured for u-boot hacking.
There are extra flash partition on the second 16MB of flash which mirror
the first:

 /dev/mtd4 - secondary u-boot
 /dev/mtd5 - secondary u-boot environment
 /dev/mtd6 - secondary "factory"
 /dev/mtd7 - secondary firmware - not used

When you 'reboot', the secondary u-boot will be run.  When you
power-cycle, the normal primary u-boot will be run.

So you need to store a u-boot in /dev/mtd4 with e.g.
  flashcp uboot.bin /dev/mtd4

and then you can test it, but still be able to recover via power-cycle.

Signed-off-by: NeilBrown <neil@brown.name>