gbmake chroot: handle dependencies better
debootstrap does not unpack all the extra packages requested until phase
two, which has to be run on the target arch.
We cannot run phase two on MIPS so we unpack them manually.
Currently we don't handle finding and unpacking the dependencies, so we
have a hard-coded list of extra things to unpack.

Add some magic to extract list of dependencies from the packages and
unpack them, so we don't need the hard-coded list.

Signed-off-by: NeilBrown <neil@brown.name>