mkinitramfs: move existence check into script.
"gbmake firmware" currently only runs mkinitramfs if it doesn't exist,
or doesn't have an 'init' file.

Change this to do the test inside the script and pass "-f" when we
explicitly want to recreate the initramfs.

As mkinitramfs also makes the files list which lives in the object
directory, it is easy for the files list to be removed while 'init'
still exists, so "gbmake firmware" might not restore the files list.

Now "gbmake firmware" will always create the files list if it doesn't
exist, and will still create rest of the initramfs if it doesn't exist.

"gbmake initramfs" will unconditionally make both.

Signed-off-by: NeilBrown <neil@brown.name>