gbmake: introduce support for appending DTB to vmlinux.bin
Upstream kernels since v5.18-rc1~81^2~7 do not support
CONFIG_BUILTIN_DTB=y for Gnubee builds.  To support these we need to
create a U-Boot bootable image ourselves, appending the chosen DTB file
to vmlinux.bin.

So introduce this as an alternate build strategy when CONFIG_BUILDIN_DTB
is not given.  In this case we always build both the pc1 and pc2 images.

Signed-off-by: Arınç ÜNAL <arinc.unal@arinc9.com>