Change trigger for building uboot-hacking version.
The "uboot-hacking" version of firmware is small enough to fit in half
of the flash - so an alternate u-boot can be tested in the second half.

So a couple of places need to decide to install less in the initramfs.

Change these tests to look at the default HOSTNAME which must now contain
the string "uboot".  The kern-config files are changed to have such a
host name.

This will allow us to remove the DTB specification from the config file
- as doing this is no longer supported upstream.

Signed-off-by: NeilBrown <neil@brown.name>